requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.0.5
django-colorfield==0.4.5
django-environ==0.7.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.3.1
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
